seperated levels, and fixed multiple events bug
